Use a default repository when working in multi-repos mode
Make a default repository to be defined as required in multi-repos.

When working with multiple repositories always have a default
repository to use in a way similar to having PACKAGESITE defined.

Installations/upgrades should be done from a single repository at a
time.

Choosing from which repo to install/upgrade is done via specifying the
repository with the '-r <repo>' flag.