[fixed] redesign of pkgdb_lock and pkgdb_unlock functions
it is assumed that lock_count==0 before lock and ==1 if the database is locked.
Asserts help us here.
Locking is implemented in a form of infinity while loop, with sleep(1) call if the database is locked by another pkgng process
EPKG_OK is the return value of the functions.
pkgdb_unlock return EPKG_FATAL in case of error while sql query.