[new_feature] pkg_fetch_file() receives open file flags as argument
it is necessary to pass flags to this function to prevent concurrent repo updates.
flags: O_WRONLY|O_CREAT|O_TRUNC|O_EXLOCK are sufficient for this purpose.
Hence process will be blocked while trying to open the file if another process is working with this file.
No need to require "non-existence" of this file, be cause O_TRUNC flag will rewrite the whole file.